Introduce ACP controller.
